# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];


export const ProductCategory = {
  Ulam: 'Ulam',
  Rice: 'Rice',
  Beverage: 'Beverage',
  'Add-on': 'Add-on',
} as const;

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  price: number;
  isAvailable: boolean;
  accent?: string;
}

export interface ProductUpdate {
  /** @minimum 0 */
  price?: number;
  isAvailable?: boolean;
}

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export type OrderInputPaymentMethod = typeof OrderInputPaymentMethod[keyof typeof OrderInputPaymentMethod];


export const OrderInputPaymentMethod = {
  Cash: 'Cash',
  GCash: 'GCash',
  Maya: 'Maya',
} as const;

export interface OrderInput {
  /** @minItems 1 */
  items: OrderItemInput[];
  paymentMethod: OrderInputPaymentMethod;
  /** @minimum 0 */
  totalAmount: number;
  /** @nullable */
  cashTendered?: number | null;
  /** @nullable */
  offlineId?: string | null;
}

export interface OrderItem {
  productId: number;
  name: string;
  quantity: number;
  unitPrice: number;
}

export type OrderPaymentMethod = typeof OrderPaymentMethod[keyof typeof OrderPaymentMethod];


export const OrderPaymentMethod = {
  Cash: 'Cash',
  GCash: 'GCash',
  Maya: 'Maya',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  completed: 'completed',
  voided: 'voided',
} as const;

export interface Order {
  id: string;
  items: OrderItem[];
  totalAmount: number;
  paymentMethod: OrderPaymentMethod;
  /** @nullable */
  cashTendered?: number | null;
  /** @nullable */
  change?: number | null;
  status: OrderStatus;
  createdAt: string;
  synced: boolean;
}

export interface DashboardSummary {
  todaySales: number;
  orderCount: number;
  averageOrder: number;
  pendingOrders: number;
  lowStockCount: number;
  expenseTotal: number;
  /** @nullable */
  topProduct?: string | null;
}

export interface ProductionLogInput {
  productId: number;
  /** @minimum 1 */
  servings: number;
  /** @nullable */
  note?: string | null;
}

export type ExpenseLogInputCategory = typeof ExpenseLogInputCategory[keyof typeof ExpenseLogInputCategory];


export const ExpenseLogInputCategory = {
  ingredients: 'ingredients',
  supplies: 'supplies',
  delivery: 'delivery',
  other: 'other',
} as const;

export interface ExpenseLogInput {
  /** @minimum 0 */
  amount: number;
  details: string;
  category?: ExpenseLogInputCategory;
}

export type OperationLogType = typeof OperationLogType[keyof typeof OperationLogType];


export const OperationLogType = {
  batch_production: 'batch_production',
  palengke_expense: 'palengke_expense',
} as const;

export interface OperationLog {
  id: string;
  type: OperationLogType;
  details: string;
  amount: number;
  createdAt: string;
}

export type ListOrdersParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};