# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ulamify POS and operations API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Menu catalog operations
  - name: orders
    description: POS order operations
  - name: operations
    description: Kitchen and purchasing operations
  - name: dashboard
    description: Owner dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List menu products
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product availability or price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a completed POS order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List recent orders
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get today's operations summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /operations/production:
    post:
      operationId: createProductionLog
      tags: [operations]
      summary: Log a kitchen batch production
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionLogInput"
      responses:
        "201":
          description: Created production log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationLog"
  /operations/expenses:
    post:
      operationId: createExpenseLog
      tags: [operations]
      summary: Log a palengke expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseLogInput"
      responses:
        "201":
          description: Created expense log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationLog"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, category, price, isAvailable]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [Ulam, Rice, Beverage, Add-on]
        price:
          type: number
        isAvailable:
          type: boolean
        accent:
          type: string
    ProductUpdate:
      type: object
      properties:
        price:
          type: number
          minimum: 0
        isAvailable:
          type: boolean
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    OrderInput:
      type: object
      required: [items, paymentMethod, totalAmount]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
        paymentMethod:
          type: string
          enum: [Cash, GCash, Maya]
        totalAmount:
          type: number
          minimum: 0
        cashTendered:
          type: ["number", "null"]
        offlineId:
          type: ["string", "null"]
    OrderItem:
      type: object
      required: [productId, name, quantity, unitPrice]
      properties:
        productId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
    Order:
      type: object
      required: [id, items, totalAmount, paymentMethod, status, createdAt, synced]
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        paymentMethod:
          type: string
          enum: [Cash, GCash, Maya]
        cashTendered:
          type: ["number", "null"]
        change:
          type: ["number", "null"]
        status:
          type: string
          enum: [completed, voided]
        createdAt:
          type: string
        synced:
          type: boolean
    DashboardSummary:
      type: object
      required: [todaySales, orderCount, averageOrder, pendingOrders, lowStockCount, expenseTotal]
      properties:
        todaySales:
          type: number
        orderCount:
          type: integer
        averageOrder:
          type: number
        pendingOrders:
          type: integer
        lowStockCount:
          type: integer
        expenseTotal:
          type: number
        topProduct:
          type: ["string", "null"]
    ProductionLogInput:
      type: object
      required: [productId, servings]
      properties:
        productId:
          type: integer
        servings:
          type: integer
          minimum: 1
        note:
          type: ["string", "null"]
    ExpenseLogInput:
      type: object
      required: [amount, details]
      properties:
        amount:
          type: number
          minimum: 0
        details:
          type: string
        category:
          type: string
          enum: [ingredients, supplies, delivery, other]
    OperationLog:
      type: object
      required: [id, type, details, amount, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [batch_production, palengke_expense]
        details:
          type: string
        amount:
          type: number
        createdAt:
          type: string