# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  todaySales: number;
  orderCount: number;
  averageOrder: number;
  pendingOrders: number;
  lowStockCount: number;
  expenseTotal: number;
  /** @nullable */
  topProduct?: string | null;
}