# lib/api-zod/src/generated/types/expenseLogInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseLogInputCategory } from './expenseLogInputCategory';

export interface ExpenseLogInput {
  /** @minimum 0 */
  amount: number;
  details: string;
  category?: ExpenseLogInputCategory;
}