# lib/api-zod/src/generated/types/operationLog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { OperationLogType } from './operationLogType';

export interface OperationLog {
  id: string;
  type: OperationLogType;
  details: string;
  amount: number;
  createdAt: string;
}