# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentMethod } from './orderPaymentMethod';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  items: OrderItem[];
  totalAmount: number;
  paymentMethod: OrderPaymentMethod;
  /** @nullable */
  cashTendered?: number | null;
  /** @nullable */
  change?: number | null;
  status: OrderStatus;
  createdAt: string;
  synced: boolean;
}