# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputPaymentMethod } from './orderInputPaymentMethod';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minItems 1 */
  items: OrderItemInput[];
  paymentMethod: OrderInputPaymentMethod;
  /** @minimum 0 */
  totalAmount: number;
  /** @nullable */
  cashTendered?: number | null;
  /** @nullable */
  offlineId?: string | null;
}