# lib/api-zod/src/generated/types/orderItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemInput {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}