# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  price: number;
  isAvailable: boolean;
  accent?: string;
}