# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minimum 0 */
  price?: number;
  isAvailable?: boolean;
}