# lib/api-zod/src/generated/types/productionLogInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Ulamify POS and operations API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductionLogInput {
  productId: number;
  /** @minimum 1 */
  servings: number;
  /** @nullable */
  note?: string | null;
}